requirements.txt
Flask
geopy
"python-socketio[asyncio_client]"
flask-socketio
pytest
